@@ -19,17 +19,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.2.2",
+    "@angular/animations": "^13.2.4",
     "@angular/cdk": "^13.1.2",
-    "@angular/common": "^13.2.3",
+    "@angular/common": "^13.2.4",
     "@angular/compiler": "^13.1.2",
     "@angular/core": "^13.1.2",
-    "@angular/forms": "^13.2.3",
-    "@angular/material": "^13.2.3",
-    "@angular/platform-browser": "^13.2.3",
-    "@angular/platform-browser-dynamic": "^13.2.3",
-    "@angular/router": "^13.2.3",
-    "@splitsoftware/splitio-browserjs": "^0.4.0",
+    "@angular/forms": "^13.2.4",
+    "@angular/material": "^13.2.4",
+    "@angular/platform-browser": "^13.2.4",
+    "@angular/platform-browser-dynamic": "^13.2.4",
+    "@angular/router": "^13.2.4",
+    "@splitsoftware/splitio-browserjs": "^0.4.1",
     "angular-gridster2": "^13.1.1",
     "ng-click-outside": "^9.0.1",
     "ngx-cookie-service": "^13.1.2",
@@ -40,37 +40,37 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1302.3",
-    "@angular-devkit/build-angular": "^13.2.4",
+    "@angular-devkit/architect": "^0.1302.5",
+    "@angular-devkit/build-angular": "^13.2.5",
     "@angular-eslint/builder": "^13.1.0",
     "@angular-eslint/eslint-plugin": "^13.1.0",
     "@angular-eslint/eslint-plugin-template": "^13.1.0",
     "@angular-eslint/schematics": "^13.1.0",
     "@angular-eslint/template-parser": "^13.1.0",
-    "@angular/cli": "^13.2.4",
-    "@angular/compiler-cli": "^13.2.3",
-    "@compodoc/compodoc": "^1.1.18",
+    "@angular/cli": "^13.2.5",
+    "@angular/compiler-cli": "^13.2.4",
+    "@compodoc/compodoc": "^1.1.19",
     "@strut-software/eslint-config": "^1.8.0",
     "@strut-software/stylelint-config": "^1.1.1",
     "@types/jasmine": "^3.10.3",
-    "@types/node": "^16.11.25",
+    "@types/node": "^16.11.26",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "5.12.0",
-    "@typescript-eslint/parser": "5.12.0",
+    "@typescript-eslint/eslint-plugin": "5.12.1",
+    "@typescript-eslint/parser": "5.12.1",
     "autoprefixer": "^10.4.2",
     "cross-env": "^7.0.3",
-    "eslint": "8.9.0",
+    "eslint": "8.10.0",
     "husky": "^7.0.4",
-    "jasmine-core": "~4.0.0",
+    "jasmine-core": "~4.0.1",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.3.16",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-mocks": "^v13.0.3",
+    "ng-mocks": "^v13.0.4",
     "pa11y-ci": "3.0.1",
-    "postcss": "^8.4.6",
+    "postcss": "^8.4.7",
     "postcss-cli": "^9.1.0",
     "prettier": "2.5.1",
     "sassdoc": "^2.7.3",
