@@ -275,6 +275,7 @@ describe('DocumentService', () => {
       workers: [],
       questions: [],
       instructions: 'General instructions',
+      isChained: false,
     };
 
     service.getDocumentInfo(stationId, documentId).subscribe((response) => {
@@ -558,16 +559,24 @@ describe('DocumentService', () => {
     httpTestingController.verify();
   });
 
-  it('should flow a document to previous station', () => {
+  it('should flow a document to previous stations', () => {
     const dataExpect: MoveDocument = {
       fromStationRithmId: stationId,
-      toStationRithmIds: ['123-654-789'],
+      toStationRithmIds: ['123-654-789', '222-654-789', '321-456-987'],
       documentRithmId: documentId,
     };
 
     service.flowDocumentToPreviousStation(dataExpect).subscribe((response) => {
       expect(response).toBeFalsy();
     });
+    const router = `${environment.baseApiUrl}${MICROSERVICE_PATH}/flow-station-to-station`;
+    const req = httpTestingController.expectOne(router);
+
+    expect(req.request.method).toEqual('POST');
+    expect(req.request.url).toEqual(router);
+    expect(req.request.body).toEqual(dataExpect);
+    req.flush(null);
+    httpTestingController.verify();
   });
 
   it('should unassign a user to document via API', () => {
