@@ -8,6 +8,9 @@ import {
   StationRosterMember,
   Document,
   DashboardData,
+  EditDataWidget,
+  ColumnsDocumentInfo,
+  ColumnsLogicWidget,
 } from 'src/models';
 
 const MICROSERVICE_PATH = '/dashboardservice/api/dashboard';
@@ -20,7 +23,38 @@ const MICROSERVICE_PATH = '/dashboardservice/api/dashboard';
 })
 export class DashboardService {
   /** Loading dashboard when generating new dashboard. */
-  isLoadingDashboard$ = new Subject<boolean>();
+  isLoadingDashboard$ = new Subject<{
+    /** To toggle loading dashboard. */
+    statusLoading: boolean;
+    /** True to load getParams in dashboard. */
+    getParams: boolean;
+  }>();
+
+  /** Update specific widget and data. */
+  updateDataWidget$ = new Subject<EditDataWidget>();
+
+  columnsDocumentInfo: ColumnsLogicWidget[] = [
+    {
+      name: 'Document',
+      key: ColumnsDocumentInfo.Name,
+    },
+    {
+      name: 'Assigned',
+      key: ColumnsDocumentInfo.AssignedUser,
+    },
+    {
+      name: 'Priority',
+      key: ColumnsDocumentInfo.Priority,
+    },
+    {
+      name: 'Time in station',
+      key: ColumnsDocumentInfo.TimeInStation,
+    },
+    {
+      name: 'Last updated',
+      key: ColumnsDocumentInfo.LastUpdated,
+    },
+  ];
 
   constructor(private http: HttpClient) {}
 
@@ -38,10 +72,20 @@ export class DashboardService {
   /**
    * Toggle emit to loading dashboard.
    *
-   * @param status Boolean true to loading and false not loading.
+   * @param statusLoading To toggle loading dashboard.
+   * @param getParams True to load getParams in dashboard.
    */
-  toggleLoadingDashboard(status: boolean): void {
-    this.isLoadingDashboard$.next(status);
+  toggleLoadingDashboard(statusLoading: boolean, getParams = false): void {
+    this.isLoadingDashboard$.next({ statusLoading, getParams });
+  }
+
+  /**
+   * Update data of the widget since drawer station.
+   *
+   * @param editDataWidget Data to edit widget.
+   */
+  updateDashboardWidgets(editDataWidget: EditDataWidget): void {
+    this.updateDataWidget$.next(editDataWidget);
   }
 
   /**
