@@ -1,7 +1,14 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import {
+  Component,
+  EventEmitter,
+  OnDestroy,
+  OnInit,
+  Output,
+} from '@angular/core';
 import { Subject } from 'rxjs';
 import { SidenavDrawerService } from 'src/app/core/sidenav-drawer.service';
 import { takeUntil } from 'rxjs/operators';
+import { PopupService } from 'src/app/core/popup.service';
 
 /**
  * Component for widget drawer.
@@ -15,10 +22,19 @@ export class WidgetDrawerComponent implements OnInit, OnDestroy {
   /** Subject for when the component is destroyed. */
   private destroyed$ = new Subject<void>();
 
-  /** Whether the called widget-drawer. */
-  drawerMode: 'stationWidget' = 'stationWidget';
+  /** Emit event for delete widget in dashboard. */
+  @Output() deleteWidget = new EventEmitter<number>();
 
-  constructor(private sidenavDrawerService: SidenavDrawerService) {}
+  /** Widget index of opened widget-drawer. */
+  widgetIndex!: number;
+
+  /** Whether the called widget-drawer. */
+  drawerMode: 'stationWidget' | 'documentWidget' = 'stationWidget';
+
+  constructor(
+    private sidenavDrawerService: SidenavDrawerService,
+    private popupService: PopupService
+  ) {}
 
   /**
    * Initial Method.
@@ -27,7 +43,7 @@ export class WidgetDrawerComponent implements OnInit, OnDestroy {
     this.sidenavDrawerService.drawerContext$
       .pipe(takeUntil(this.destroyed$))
       .subscribe((data) => {
-        if (data === 'stationWidget') {
+        if (data === 'stationWidget' || data === 'documentWidget') {
           this.drawerMode = data;
         }
       });
@@ -38,9 +54,39 @@ export class WidgetDrawerComponent implements OnInit, OnDestroy {
    *
    */
   toggleDrawer(): void {
-    if (this.drawerMode === 'stationWidget') {
+    if (
+      this.drawerMode === 'stationWidget' ||
+      this.drawerMode === 'documentWidget'
+    ) {
       this.sidenavDrawerService.toggleDrawer(this.drawerMode);
     }
+  }
+
+  /**
+   * Show alert confirm widget delete.
+   *
+   */
+  async confirmWidgetDelete(): Promise<void> {
+    const response = await this.popupService.confirm({
+      title: 'Delete Widget?',
+      message: 'This cannot be undone!',
+      okButtonText: 'Yes',
+      cancelButtonText: 'No',
+      important: true,
+    });
+    if (response) {
+      this.toggleDrawer();
+      this.deleteWidget.emit(this.widgetIndex);
+    }
+  }
+
+  /**
+   * Event emit widgetIndex to dashboard.
+   *
+   * @param widgetIndex Widget index from station-widget-drawer.
+   */
+  setWidgetIndex(widgetIndex: number): void {
+    this.widgetIndex = widgetIndex;
   }
 
   /**
