@@ -1,23 +1,35 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { first } from 'rxjs';
+import {
+  Component,
+  Input,
+  OnInit,
+  OnDestroy,
+  Output,
+  EventEmitter,
+} from '@angular/core';
+import { first, Subject } from 'rxjs';
 import { DocumentService } from 'src/app/core/document.service';
 import { ErrorService } from 'src/app/core/error.service';
-import { DocumentWidget } from 'src/models';
+import { DocumentWidget, QuestionFieldType } from 'src/models';
 import { Router } from '@angular/router';
+import { SidenavDrawerService } from 'src/app/core/sidenav-drawer.service';
+import { takeUntil } from 'rxjs/operators';
 
 /**
  * Component for list field the document how widget.
  */
 @Component({
-  selector: 'app-document-widget[documentRithmId]',
+  selector: 'app-document-widget[documentRithmId][editMode]',
   templateUrl: './document-widget.component.html',
   styleUrls: ['./document-widget.component.scss'],
 })
-export class DocumentWidgetComponent implements OnInit {
+export class DocumentWidgetComponent implements OnInit, OnDestroy {
   /** Document rithmId. */
   @Input() documentRithmId = '';
 
-  /** Data to document list for widget. */
+  /** Edit mode toggle from dashboard. */
+  @Input() editMode = false;
+
+  /** Data document list for widget. */
   dataDocumentWidget!: DocumentWidget;
 
   /** Loading document widget. */
@@ -26,11 +38,32 @@ export class DocumentWidgetComponent implements OnInit {
   /** Show error if get documentWidget fails. */
   failedLoadWidget = false;
 
+  /** The question field type. */
+  questionFieldType = QuestionFieldType;
+
+  /** Type of drawer opened. */
+  drawerContext!: string;
+
+  /** Open drawer. */
+  @Output() toggleDrawer = new EventEmitter<void>();
+
+  private destroyed$ = new Subject<void>();
+
   constructor(
     private errorService: ErrorService,
     private documentService: DocumentService,
-    private router: Router
+    private router: Router,
+    private sidenavDrawerService: SidenavDrawerService
   ) {}
+
+  /**
+   * Whether the drawer is open.
+   *
+   * @returns True if the drawer is open, false otherwise.
+   */
+  get isDrawerOpen(): boolean {
+    return this.sidenavDrawerService.isDrawerOpen;
+  }
 
   /**
    * Initial Method.
@@ -38,6 +71,11 @@ export class DocumentWidgetComponent implements OnInit {
   ngOnInit(): void {
     this.documentRithmId = JSON.parse(this.documentRithmId).documentRithmId;
     this.getDocumentWidget();
+    this.sidenavDrawerService.drawerContext$
+      .pipe(takeUntil(this.destroyed$))
+      .subscribe((drawerContext) => {
+        this.drawerContext = drawerContext;
+      });
   }
 
   /**
@@ -66,7 +104,8 @@ export class DocumentWidgetComponent implements OnInit {
       });
   }
 
-  /** Navigate the user to the document page.
+  /**
+   * Navigate the user to the document page.
    *
    * @param stationId The Id of the station in which to view the document.
    */
@@ -77,5 +116,16 @@ export class DocumentWidgetComponent implements OnInit {
         stationId: stationId,
       },
     });
+  }
+
+  /** Toggle drawer when click on edit station widget. */
+  toggleEditDocument(): void {
+    this.toggleDrawer.emit();
+  }
+
+  /** Clean subscriptions. */
+  ngOnDestroy(): void {
+    this.destroyed$.next();
+    this.destroyed$.complete();
   }
 }
