@@ -6,6 +6,8 @@ import {
   WidgetType,
   DashboardData,
   RoleDashboardMenu,
+  EditDataWidget,
+  ColumnsDocumentInfo,
 } from 'src/models';
 import { delay } from 'rxjs/operators';
 import { Document } from 'src/models';
@@ -16,16 +18,62 @@ import { Document } from 'src/models';
  * Mocks methods of the `DashboardService`.
  */
 export class MockDashboardService {
-  /** Loading dashboard when generate new dashboard. */
-  isLoadingDashboard$ = new Subject<boolean>();
+  /** Loading dashboard when generating new dashboard. */
+  isLoadingDashboard$ = new Subject<{
+    /** To toggle loading dashboard. */
+    statusLoading: boolean;
+    /** True to load getParams in dashboard. */
+    getParams: boolean;
+  }>();
+
+  /** Update specific widget and data. */
+  updateDataWidget$ = new Subject<EditDataWidget>();
+
+  columnsDocumentInfo: {
+    /** Name to show in dom. */
+    name: string;
+    /** Key property to get value of the document. */
+    key: string;
+  }[] = [
+    {
+      name: 'Document',
+      key: ColumnsDocumentInfo.Name,
+    },
+    {
+      name: 'Assigned',
+      key: ColumnsDocumentInfo.AssignedUser,
+    },
+    {
+      name: 'Priority',
+      key: ColumnsDocumentInfo.Priority,
+    },
+    {
+      name: 'Time in station',
+      key: ColumnsDocumentInfo.TimeInStation,
+    },
+    {
+      name: 'Last updated',
+      key: ColumnsDocumentInfo.LastUpdated,
+    },
+  ];
+
+  /**
+   * Update data of the widget since drawer station.
+   *
+   * @param editDataWidget Data to edit widget.
+   */
+  updateDashboardWidgets(editDataWidget: EditDataWidget): void {
+    this.updateDataWidget$.next(editDataWidget);
+  }
 
   /**
    * Toggle emit to loading dashboard.
    *
-   * @param status Boolean true to loading and false not loading.
+   * @param statusLoading To toggle loading dashboard.
+   * @param getParams True to load getParams in dashboard.
    */
-  toggleLoadingDashboard(status: boolean): void {
-    this.isLoadingDashboard$.next(status);
+  toggleLoadingDashboard(statusLoading: boolean, getParams = false): void {
+    this.isLoadingDashboard$.next({ statusLoading, getParams });
   }
 
   /**
